requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.4.0
# gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
#boto==2.49.0
boto3==1.9.146
django-storages==1.7.1


# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==6.0

# Raven is the Sentry client
# --------------------------
raven==6.10.0

# ubuntu package is old and has issues, install from pip instead
mod_wsgi==4.6.5

